fix: kiểm tra quyền thao tác CRUD admin

+++ Shop/Areas/Administrator/Controllers/AspNetRolesController.cs
@@ -17,28 +17,49 @@ namespace Shop.Areas.Administrator.Controllers
         // GET: Administrator/AspNetRoles
         public ActionResult Index()
         {
-            return View(db.AspNetRoles.ToList());
+            if (Session["taikhoanadmin"] == null)
+            {
+                return RedirectToAction("Error401", "MainPage");
+            }
+            else
+            {
+                return View(db.AspNetRoles.ToList());
+            }
         }
 
         // GET: Administrator/AspNetRoles/Details/5
         public ActionResult Details(string id)
         {
-            if (id == null)
+            if (Session["taikhoanadmin"] == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Error401", "MainPage");
             }
-            AspNetRole aspNetRole = db.AspNetRoles.Find(id);
-            if (aspNetRole == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                AspNetRole aspNetRole = db.AspNetRoles.Find(id);
+                if (aspNetRole == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(aspNetRole);
             }
-            return View(aspNetRole);
         }
 
         // GET: Administrator/AspNetRoles/Create
         public ActionResult Create()
         {
-            return View();
+            if (Session["taikhoanadmin"] == null)
+            {
+                return RedirectToAction("Error401", "MainPage");
+            }
+            else
+            {
+                return View();
+            }
         }
 
         // POST: Administrator/AspNetRoles/Create
@@ -48,29 +69,43 @@ namespace Shop.Areas.Administrator.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,Name")] AspNetRole aspNetRole)
         {
-            if (ModelState.IsValid)
+            if (Session["taikhoanadmin"] == null)
             {
-                db.AspNetRoles.Add(aspNetRole);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Error401", "MainPage");
             }
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    db.AspNetRoles.Add(aspNetRole);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
 
-            return View(aspNetRole);
+                return View(aspNetRole);
+            }
         }
 
         // GET: Administrator/AspNetRoles/Edit/5
         public ActionResult Edit(string id)
         {
-            if (id == null)
+            if (Session["taikhoanadmin"] == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Error401", "MainPage");
             }
-            AspNetRole aspNetRole = db.AspNetRoles.Find(id);
-            if (aspNetRole == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                AspNetRole aspNetRole = db.AspNetRoles.Find(id);
+                if (aspNetRole == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(aspNetRole);
             }
-            return View(aspNetRole);
         }
 
         // POST: Administrator/AspNetRoles/Edit/5
@@ -80,28 +115,42 @@ namespace Shop.Areas.Administrator.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,Name")] AspNetRole aspNetRole)
         {
-            if (ModelState.IsValid)
+            if (Session["taikhoanadmin"] == null)
             {
-                db.Entry(aspNetRole).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Error401", "MainPage");
             }
-            return View(aspNetRole);
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    db.Entry(aspNetRole).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View(aspNetRole);
+            }
         }
 
         // GET: Administrator/AspNetRoles/Delete/5
         public ActionResult Delete(string id)
         {
-            if (id == null)
+            if (Session["taikhoanadmin"] == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Error401", "MainPage");
             }
-            AspNetRole aspNetRole = db.AspNetRoles.Find(id);
-            if (aspNetRole == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                AspNetRole aspNetRole = db.AspNetRoles.Find(id);
+                if (aspNetRole == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(aspNetRole);
             }
-            return View(aspNetRole);
         }
 
         // POST: Administrator/AspNetRoles/Delete/5
@@ -109,10 +158,17 @@ namespace Shop.Areas.Administrator.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(string id)
         {
-            AspNetRole aspNetRole = db.AspNetRoles.Find(id);
-            db.AspNetRoles.Remove(aspNetRole);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Session["taikhoanadmin"] == null)
+            {
+                return RedirectToAction("Error401", "MainPage");
+            }
+            else
+            {
+                AspNetRole aspNetRole = db.AspNetRoles.Find(id);
+                db.AspNetRoles.Remove(aspNetRole);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
         }
 
         protected override void Dispose(bool disposing)
